pyproject.toml
[tool.poetry]
name = "imeicheckbot"
version = "0.1.0"
description = ""
authors = ["md <marceldiomare@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.37"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
cryptography = "^44.0.0"
aiogram = "^3.17.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
alembic = "^1.14.1"
cloudscraper = "^1.2.71"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
